CronHelper: Don't bother with current time
Time is a fickle thing.

This patch defines a predictable point in time, from which we deduce our
intervals instead of messing about with the local system time, which is
affected by leap years, and all that scary stuff.

In addition, this patch fixes a bug where we would accidentally get
twice the interval length, if the method was run at the execution time
of the supplied cron (just typing that makes me feel dirty).