Handle empty gateway (and meter) fields
Prior to this patch, we would try to save a gateway without a serial,
which is not allowed by the database schema. This patch avoids that
issue, by doing further validation of the gateway before trying to
create it.

Empty meter fields were already handled properly, but this patch adds a
couple of tests to make sure they stay that way.