mq: Require meter ID for creating physical meters
All physical meters should already have an ID, so we cannot create it on
demand.

I looked into invalidating the MeterDto if its ID was invalid, but I
could not do it in a pretty way.

We could add it to the GSON parser, which would require us to special
case every class and would be spammy (much like this intermediate bug
fix is). We could set the field on inspection when traversing through
the object graph but that would require setting values based on
reflection. This is the easier solution, until something smarter bubbles
up.