gitlab: Update salt earlier in release pipeline
The "upgrade salt filesystem" event is async, so we must wait for the
"publish:update" step to finish. We're not sure how to do that though,
so instead, we fire the event earlier, in the prepare stage.

The salt job depends on the ci-utils docker image, so we move the
creation of all docker images back a bit.