Introduce properties use case to force update geo location for an entity
The force parameter is sent along the request parameters to geo service
now when sync button is clicked from evo.

Note that 'update.geolocation' property is read and then when the request
is sent to geoservice, the  property is removed. But only this property
for now. Later we might have properties that are enabled/disabled such as
feature toggling, and therefore the name of the enum is `FeatureType`.