Disable auditing for property table
This is just temporary table actually, don't know if it needs to be
audited. However, there seams to be a bug somewhere with hibernate envers.
It just keeps throwing exceptions in an infinite loop when syncing a meter
with metering a second time!

Addresses issue: https://elvaco.atlassian.net/browse/MVP-541