Increase throttling cache size
In a worst case scenario, we're missing information for every meter in
the system (such as, when we realize that we're missing a
report_interval for many meters). A low number of allowed entries
in the throttling heap would cause older entries to be evicted
before we're done with the current burst, causing plenty of ref-info
message duplicates, unnecessarily.