Zoom in on meter in dashboard view
Even meters missing a geoposition get a zoom-icon which sadly is a
silent noop. This may confuse users. It is left this way, because the
current structure of the selection tree (ST) does not include
geoposition, and the ST is the subject of an upcoming refactoring.

We now show the ST in all views. I felt that showing the ST in 50% of
the views is visually annoying when navigating quickly, it flickers too
much.

This fixes MVP-555