Add txSize and prefetchCount to message listener container
Defaulted to 100 message to be delivered in a transaction to reduce the
number of acks. We found this to be a breaking point on my machine.

Regarding the prefetchCount this is a copy-paste from the documentation,
and the reason why we added this property to be >= txSize.

"The number of unacknowledged messages that can be outstanding at each
consumer. The higher this is the faster the messages can be delivered,
but the higher the risk of non-sequential processing"

Note that non-sequential processing is not problem for us, since the do
not consume messages in any specific sequence.