Missing measurements view, performance tuning
This patch should reduce the time required to refresh the
'missing_measurements' view from ~6hrs to ~25 minutes, while still
keeping it compatible with most of the existing code (barring stuff that
we never used or needed anyways, which I've removed).

The only functional change is that missing measurements are now
calculated up until the end of the current day (`date_trunc('hour',
current_timestamp)`).