selection-tree: Do not cause table to reload
Searching the selection tree still stores the state in SearchState
alongside the collection and validation search queries. Re-using the
SEARCH action was a mistake though, that caused far too many reducers to
listen in on it, and react to the changes by reloading the  list tables
in the collection & validation views.

Solution: create a new action for the selection tree.