report: Hide average from table legend
Average lines are automagically rendered, when multiple meters are
selected. They are removed if the user deselects meters until there are
0 or 1 left, in which case there is no reason to render an average.

In order to solve this, we add a new property on LineProps: "origin",
which corresponds to one of the methods in the measurement controller:
measurement (individual meters), average, cities.