Emit jobIds for synchronization requests
This patch tags synchronization requests with a random UUID as the job
ID, and returns a mapping between these job IDs and the respective
logical meter ID in the response body to the user who triggered the
request.

This'll allow us to track a request and its eventual response when/if it
arrives in the future.