medium: Always apply in normalize, remove dead media
There is no "inside temperature" in the back end, so it was replaced in
the front end with "room sensor" which is an actual thing.

Passing medium as strings everywhere is ugly. We normalize them to
Medium right away instead, enabling us to use the correct types
everywhere and noticing bugs that were easy to fix:
- Medium enum had unused "collection" and "measurementQuality" only
  "used" by widgets, but the rendering of widgets didn't even look at
  its "type" property (which supposedly could carry the value of e.g.
  "collection"), so that was removed
- Some media could not be rendered in report, because they weren't
  mapped to any quantities
- The icon component errored out, because there was a non-existing
  medium passed all the way down. This is now secured by always
  assigning a medium, and double secured with a default icon for all
  media not defined, to get an IconUnknown icon
- Inside and temperature were removed, no meters exist with those meter
  definitions