meter-dialog: Handle missing quantities
A recent rewrite (790c9ade8) ordered the quantities in the meter dialog.
When no longer extracting the available columns from the API response,
but looking at a prebuilt list of quantities, caused the meter dialog to
blow up for meter's with "incomplete" quantities in comparision to their
meter definition.

This fixes MVP-645.