setSelectedEntries() requires more than IDs
In preparation for sending a larger payload, we replace the payload
uuid[] with {ids: uuid[]}. That object will soon get more attributes, to
support selecting default medium indicators and quantities, within a
single "select this meter" action.

We also add types to our payload and it's action, so that we get type
errors instead of test errors, which is very helpful.