MeasurementController: Make /cities accept different media
There was a tight coupling between the input city and quantities, and
the meters that actually existed in the city. Previously, there was an
error if we requested a valid quantity but the city only had meters for
a different quantity. Some more errors of that kind were also corrected
and got their own tests.