ProfileContainer: Do not send argument to Logout
Åsa noticed a "createPersistoid()" error, which belonged to
redux-persist, which was thrown when trying to serialize the current
state, when the state contained non-serializable data.

When calling the action logout(error), the error is supposed to be
something returned from the API, like "not logged in".

The way ProfileContainer was coded, it called logout(event) where event
was a click-event. That's wrong, and the cause of the bug. The solution
is easy: call logout() without any arguments.

A small style issue is also fixed: wrapping things in a React.Fragment
is much easier to read than keeping elements in an array.