Adjust existing measurement timestampts to correct offset
Since we accidentally misinterpreted incoming measurements from Metering
as originating in CEST, we now need to adjust them back to the correct offset.

Note that we have already lost one hour of measurements for all meters
that were active at the time, when DST was last observed.

This also reapplies:
e2662de3f "Use timezone UTC+1 for metering messages as CET calculates
with daylight saving"

Since this is now safe.

This fixes MVP-757 (and MVP-755, again)