Consumption values should be shown on start time of each report interval, e.g the consumption for 00-01 should be on hour 00 in the report page.
The from and to should be inclusive,
e.g. asking for 00 to 00 next day should give 25 hours.

If there are gaps in the measurement the consumption for the gap should be presented
at the time before the gap.

No future measurements should be used to calculate the last consumption value if the next
one is missing.