Organisation: Dispatch post to custom url
Constructing the URL /api/v1/organisations/{id}/sub-organisations was
not possible with postRequest, so we add postRequestToUrl which takes an
url-formatting callback.

Sadly, the types are "any" for now, since I couldn't get the callback
(hell) to match what I wanted, probably because of Redux modifying the
functions through bindActionCreators().