Remove Serializable from domain models
Serializable was added for the sake of cacheability. However
- ehcache only requires serializability for storing cached elements to
disk; something that we don't currently make use of.

Quoting the ehcache FAQ:

"Can non-Serializable objects be stored in a cache?

As of Ehcache 1.2, they can be stored in caches with MemoryStores. If an
attempt is made to replicate or overflow a non-serializable element to
disk, the element is removed and a warning is logged."
(http://www.ehcache.org/documentation/2.8/faq.html)

In addition, the ConcurrentMapCache we typically use for domain models
returned from Repositories, also doesn't require a Serializable
contract, as long as "store-by-value" is disabled (the default, which we
use).