Consolidate application configuration properties
The properties for this prefix were spread out all over the place. This
puts them all together, and represents them with a
ConfigurationProperties class, to enable neat autocompletion for all of
them.

I intentionally left consumer/Rabbit properties out of this, since
there's quite a lot of them, they're isolated, and they work fine with
the autocompletion as is.