Sync measurements with average in the report graph
The measurements for a period and resolution should include all expected measurement
timestamps with null as value where the measurement is missing.
From and to arguments are considered as inclusive.
Only values at the resolution points are included, not any values in between.
For example measurements on minutes in between the whole hour for hour resolution
are excluded.

The average is based on the values shown in the graph, e.g any measurements in
between the resoultion points are excluded from the average calculation.