OrganisationEditForm: Set default selection in state
If a user picked a parent organisation, there would be a default option
selected in the "selection dropdown". That option was not added to the
state though, so when the form got submitted, there was only a parentId
but not a selectionId.

Now, we set the selectionId when selecting a parent, as long as we're
not selecting the "no parent organisation" option.