clusterHelper: Anchor single map icons at bottom-center
When zooming in and out, the map icon would appear to jump all over the
place. It really depends on your perspective though. The icon is a pin
marker, that is "nailed" to the map at the bottom & center of the icon.
Not all icons look like that, so the map rendering library did not know
which point of the icon's edge, to pin at the map.

This solves MVP-811.