tests: Fix a DST sensitive test case
When run the day after shifting back from DST on a JVM observing
such rules, 'collectionStatusMeterChangeWithIntervalUpdate' would report
a slightly better collection percentage than would otherwise be
expected.

This is awful - our tests shouldn't rely on the runtime time zone.
Instead, fix the fixture at a point in time which is not sensitive to
DST.