Salt: Always fetch newest container image
When we try to deploy a develop build we don't fetch the most
recent image from our docker registry, this is how docker works.

So we need to force fetch always to get the latest image with tag
development.

When it comes to a tagged build it will need to download the image
anyway, so theres no difference in the logic.

Signed-off-by: Mattias Ryrlén <mattias.ryrlen@elvaco.se>