Rename TableInfoText to TimestampInfoMessage
TableInfoText sounds like it should be generic, but the component is
very specific - it includes a hard coded message.

Also, when looking at other components that use the caveat, it's easier
to understand exactly what it contains. If we need something more
generic in the future, we can add that then.