Introduce gateway summary DTO
We used to fetch an amalgamation of projections and entities here, which
caused N+1 behaviour as the mappers started traversing relationships.

This change introduces a 'dto' package into core, which is intended to
hold thin representations of our domain models, to avoid having to do
mapping in several steps "just because".