mapHelper: Set bounds even for a single meter
We used to (unknowingly) require more than one meter to calculate the
bounds of all markers to render on a map. Requiring only a single marker
is good enough, so now we allow that.

Before this patch, having a single meter selected and rendering the map
on the dashboard, would not center the map on that marker.

This fixes MVP-823