Replace unit_at() with application layer logic
We use UomUnitConverter to perform the unit translation.

The storage_unit is used from the entity layer and down, but known by
the domain model too. If it is null there, the entitymapper looks up the
correct unit, keeping the input and output units separated.

The MeasurementEntity now has a `double value` instead of a
MeasurementUnit. MeasurementUnit is mostly replaced by the entitymapper
+ the unit converter.