Add measurement stat table
This table aggregates various statistics about meters and their
measurements, to improve the performance of common queries.

This patch also adds time zone information to logical meters, so that we
may decide for a given meter when a "day" starts, and ends (and
subsequently aggregate date for that "day").

This fixes MVP-852.