Measurements: meter -> logicalMeterId, quantities -> quantity
This conforms with our RequestParameters.

Since we now ask the /measurements/average endpoint about city averages,
there were some changes in the tests that assumed that "a single city"
is the focal point of the whole request, which it is not any longer.
That is, an average request can include data from multiple cities (or
any other combination of selection criteria for that matter).