Move resposibility of all medium mappings to MeteringMessageMapper
PhyscalMeter contains the original medium string sent from Metering, this value is
never used.
LogicalMeter contains MeterDefiniton with a medium mapped to Medium enum used in Evo.
Metering specific mappings are removed from Medium enum.
ROOM_TEMP is renamed to ROOM_SENSOR as temp is only one of the quantities for the medium.