DI quantities, roles, meter definitions
The RabbitMQ-listeners are started before the ProductionDataLoader is
done saving all fixtures.

When starting the application against a queue
that contains measurement messages, the application would throw out
valid messages because it had no valid meter definitions/quantities to
match against the message.

The solution to this is to actually inject dependencies, instead of
having a separate ProductionDataLoader that does things.

Because of quantities having auto-IDs, we need to check for them before
saving them, and to avoid a circual dependency, we create a minimal set
of repository/JPA-repository to support looking up & saving the initial
quantities.