Slight summary count speedup
Instead of counting distinct(*), we now count what we're actually
interested in for meters (distinct IDs), which is both more correct and
faster.

Also, we now do what results in a 'select count(distinct ...)' instead
of the previous implementation which did 'select count(*) from (select
distinct ...)'. Should not make a huge difference, performance-wise, but
the SQL looks neater.