ci: Run requests against staging, confirming basic performance
The specified user needs to exist on staging for this to work, and it
should be connected to an organisation with a large amount of everything
(especially meters).

Heads up: we're using python3 async/await to accomplish parallel
requests.

As a bonus, we make the previous step wait for application to startup
before continuing with testing.