UomUnitConverter: Do not sort map for each convert()
We call convert() every time we read a measurement message. It is really
ugly to sort the map's keys based on key length every time, now we just
do it when we initialize the static map.

The test covers the algorithm's logic (don't look for and replace "m3"
before "m3/h", because then the latter would never be replaced), so we
have nothing to be afraid of.