Remove stale rabbitmq docs
We neither have "get alarm" nor "get measurements" from Metering.

The "get reference info" example is good to go, but can also be looked
up through the tests. If we wanted to have a complete set of (untested?)
messages in the docs/, we should include the messages we listen to as
well. For now, we seem to do fine by just looking at what's in the
tests.

Having these messages, that are not standardized/tested, was good for
visualizing the implementation but can be dangerous to have lying
around.