checkstyle: Use 'SAME_PACKAGE(3)' as a separate group
We don't want to interleave imports from com.elvaco.mvp with imports
from co.elastic, etc. This defines that a imports from the same package
root should be in a separate group.

Unfortunately, since we also want separateLineBetweenGroups to be true,
we have to reformat all our imports.

Coming up!