Don't handle "unhandled" exceptions
Having a catch-all exception handler means that APM cannot capture
unexpected errors and report them. That's not ideal.

This patch instead adds an ErrorAttributes implementation to format
internal errors the way we want them, while also retaining the throw.

For testing purposes, this patch also adds a new API endpoint
/api/v1/error which always triggers an internal server error. Could also
be useful for monitoring, if you're into that.