Use UnauthorizedDto explicitly, not implicitly
Using sendError() yields control over error mapping to ErrorAttributes.
This caused Unauthorized errors to be mapped as Internal Server Errors,
which is wrong. We've historically only been able to get away with this
because our UnauthorizedDto was retrofitted to conform to the format
rendered by the DefaultErrorAttribute implementation in Spring.

This makes the contract more explicit.