Bug-fix: Remove duplicate logical meters
Duplicate logical meters are removed in the application layer, here in
java, with a Set<E> usage instead of a List<E>. This happens when a
logical meter have several physical meters.

Note that, these meters are not filtered out with a distinct sql
query right now, but will be filtered out later when we know exactly
which  physical meter that is active within given period.

Resolves: https://elvaco.atlassian.net/browse/MVP-909