Let Spring handle ErrorAttributes again
This was more involved than I'd hoped. It's easier for us to simply
accept that the "Internal Server Error" message is "Internal Server
Error" (rather than "<...>, please contact support", than it is to
handle all of the different edge cases that pop up if we want to
customize it.

This makes Spring-Security originating errors appear as they used
to do again, while still allowing us to collect error information for
APM.