Re-wire test fixtures through new ContextDsl interface
This allows us to get rid of a lot almost-but-not-quite duplicated
fixture construction code and use a common way to get, for example a
logical meter, and some measurements.

The fixtures provide you with pretty sane defaults, which allows you to
simply override the stuff that's interesting for the test you're trying
to write.

This fixes MVP-916.