deploy: Fail on nonzero retcodes
In the scenario our salt-master dying and an entire deploy
catastrophically failing, we would previously celebrate and consider
everything a great success and say nothing to the contrary.

That's the default behaviour in pepper.

This new option should cause deploys to fail when they fail instead, so
we can be notified of their failure and do something about it.