dialog: Display gateway's IP and phone number
The gateway domain model already had the fields for IP and phone number
but the DTOs did not. I modified the "mandatory DTO", since that's the
one used in this particular case (the meter dialog).

The .po files were modified a lot, since it was a while ago I used
POEdit to edit them, and that program finds strings no longer in use,
and really reorders the whole file.. so that's nice.

GatewayMandatoryDto now has a @Builder.