redux: Clear state when logging in as new user
There's a new entry in the rootstate - previousSession - which currently
consists of an optional "previously logged in as user"'s ID. If we're
logging in as the same user, do nothing; if we're logging in as a
different user, dispatch `RESET_SELECTION`.

There was a problem dispatching `resetSelection()` within the `login()`
action dispatch. `yarn test` complained about not being able to find
`clearErrors()`, which is weird. Replacing the imported, typed action
with a plain object did wonders - thanks Must!

This fixes MVP-971