Add index for quick sorting of meters by medium
I remembered to define an index from type -> medium for fast lookup of
medium name by type, but forgot to define the (much more important)
inverse index, for when we want to know the order of medium names and
filter meters based on that.

This should significantly speed up sorting by medium in the meter list.