Don't offset expected measurements start by one interval
This was likely introduced due to the issue fixed by the previous
commit. In practice, it makes little sense to throw this away. It
becomes awfully obvious that something's wrong once you look at a meter
with a 24h-report interval, since that causes the "missing measurements"
to start at the day before, despite that day not being included in the
period.

This fixes MVP-1087.