Improve /average and /measurements query performance
We do this by avoiding a lot of pre-fetching and pre-processing of the
provided meters. Instead we send the entire RequestParameters to the
repository, which allows us to perform meter filtering there instead,
just like we do for regular meter queries.

Don't generate a series to match against, that takes a lot of time.
Nudge returned series in Java instead.