rabbit: Consume concurrently
This patch increases the max-concurrency for our consumer containers to
4 (pretty arbitrarily chosen, but matches the number of cores on staging
& prod).

To allow for some amount of raciness in the creation of new objects
(meters,gateways,organisations,...) we also add a fairly generous retry
policy and backoff policy to avoid DLX:ing messages unnecessarily at the
cost of retrying invalid messages unnecessarily. This should not be a
problem in practice since our DLQ is generally close-to-empty.

This fixes MVP-832.