Hide Collection % from meter page
The meaning of this is pretty confusing to the user. It is actually the
collection percentage for the selected period of the retrieved meter,
but since this period is implicit, there's no way for the user to know
this.

Removing this at @asagod's request, until we've decided the
how, if, what & when of individual meters' collection percentage in the
meter page.