Move component code from containers
Also changed the types for `OnClick` and on dom event clicks that are
just called `OnClickEventHandler`. Please see the changes in Types.ts.
This is because some clicks are callback-functions without any arguments
 that dispatches another callback function from a higher level with
 input arguments. This change separates them!

There is more to do with some of the usages of `OnClickEventHandler`.
But I leave it as it now. New code is using the properly.