Create alarm mapping table
A completely static table to start with. The data is based on an export
from Metering (typos and all!). I've excluded rows where the description
was null (because that doesn't do us any good, now does it?).

I also took the liberty of removing the description column from
meter_alarm_log since it was always null in practice, and I feel like
it's better not to duplicate the description, but rather keep it in one
canonical place in the database.