Make IDEA not require an output dir for integrationTest
The default for some reason causes IDEA to add the generated sources
directory as a source dir for the integrationTest module, which causes
it to also require an output directory, which makes no sense (since it's
a test source set)

For some reason, this is not a problem for the default test source set,
but I haven't dug enough into the idea-plugin code to figure out what
exactly causes this unfortunate behaviour. Not even sure if it's a bug,
but it sure is undesirable.