Update active period when receiving measurement before start of current active period
Period is never updated to a later timestamp.

If there is a another meter that was active at the timestamp of the measurements it's
period will be updated to stop att the timestamp. The result of this can be that existing
measurements of the previous meter is outside it's active period but this align with the
current behaviour of closing current active period when a new meter appear.

If there are two meters trying to be active at the same time the database constraint will
prevent overlapping period and this incorrect setup needs to be handled manually.