backend: Support for uploading logo/login-bg/login-logo
We serve image assets through a non-authed endpoint looking like
/organisations/:id/assets/:asset_type

That endpoint just dumps a binary blob with the uploaded content-type
(one of the allowed image types). If the organisation does not exist, we
fallback to a hardcoded Elvaco-themed asset (looking like pre-this patch).

Super admin's, or an organisation's admin, can upload these assets.