frontend: Support custom logotypes/login-bg/login-logotype
We stop referring to logged in some asset files, and instead request
those assets through the API.

The organisation administration page is somewhat revamped, to make place
for adding configuring (image) assets.

There are new actions that PUT/DELETE against URLs looking like
/organisations/:id/assets/:asset_id

We now display a link to the custom login page for an organisation, for
convenience.