Cache assets, using ETag
We send fewer bytes but we still hit the DB and calculate a string from
each image's content. Future improvement: put this in the database and
calculate the ETag on creation.

I *think* MD5 over images' byte data disqualify us from having to
prepend W/ ("weak") to our ETag. Collisions in this situation should be
crafted, and the impact should be close to zero.

In the front end, we apply a stupid "cache bust" URL parameter to the
images being previewed, in order to always force a new request
(`forceUpdateHook()` doesn't really help us here, since the props are
OK, but the images needs to be re-requested).