custom-logo: Fallback to default, not to Elvaco
If Elvaco changes its assets, all other organisations will receive the
Elvaco's images. We want to separate these two subjects, and rely on
deploying a new application artifact for changing the default assets.

Consider the case where evo is hosted on-premises of a customer, and the
front end explicitly requests Elvaco's logotypes. Yuck!