Do not cache empty optionals for entities that will be created when missing
Race condition appear when two threads try to create an organisation and
the second thread caches empty optional in between save (cache evict)
and commit of the first thread.

Sync = true is not allowed together with unless but was added trying to
solve the same problem so should be safe to replace with unless.