Ensure organisation id for super admin
This makes sure that super admin does not make a request without an
organisation id set on RequestParameters. The user's organisation id is
enforces when organisationId is not present on the url.

Selections api is the only use case that still allows super admin user
to make request without organisationId parameter.